Share some parts between Deno / Node Sqlite drivers
